:art: move `tenant_api_key` method to auth module
